+++ beaconChain/src/constants/enums.ts
