+++ beaconChain/src/constants/index.ts
