+++ beaconChain/src/helpers/validatorStatus.ts
