+++ beaconChain/src/pow/fake.ts
