+++ beaconChain/src/stateTransition/block/attestations.ts
@@ -8,7 +8,7 @@ import {
   BeaconState,
   Crosslink,
   PendingAttestation,
-} from "../../../types";
+} from "../../types";
 
 import {
   Domain,
@@ -16,7 +16,7 @@ import {
   MIN_ATTESTATION_INCLUSION_DELAY,
   SLOTS_PER_EPOCH,
   ZERO_HASH,
-} from "../../../constants";
+} from "../../constants";
 
 import {
   getAttestationParticipants,
@@ -27,7 +27,7 @@ import {
   getDomain,
   getEpochStartSlot,
   slotToEpoch,
-} from "../../../helpers/stateTransitionHelpers";
+} from "../../helpers/stateTransitionHelpers";
 
 import {blsAggregatePubkeys, blsVerifyMultiple} from "../../stubs/bls";
