+++ beaconChain/src/stateTransition/block/deposits.ts
@@ -6,17 +6,17 @@ import {
   BeaconBlock,
   BeaconState,
   DepositData,
-} from "../../../types";
+} from "../../types";
 
 import {
   DEPOSIT_CONTRACT_TREE_DEPTH,
   MAX_DEPOSITS,
-} from "../../../constants";
+} from "../../constants";
 
 import {
   hash,
   verifyMerkleBranch,
-} from "../../../helpers/stateTransitionHelpers";
+} from "../../helpers/stateTransitionHelpers";
 
 import { processDeposit} from "../../state";
