+++ beaconChain/src/types/attestation.ts
