+++ beaconChain/src/types/eth1.ts
