+++ beaconChain/src/types/index.ts
