+++ beaconChain/src/types/primitive.ts
