+++ beaconChain/src/types/state.ts
